fix: deprecate scout-url command, move helpers to cmd_url_inspect
